+++ assets/js/admin.js
@@ -1,1 +1,0 @@
-CKEDITOR.replace('page_content');