+++ assets/js/app.js
@@ -1,97 +1,22 @@
-import 'bootstrap/dist/js/bootstrap.bundle';
-import 'superfish/dist/js/superfish'
-import 'jquery-sticky/jquery.sticky'
-import 'jquery.easing/jquery.easing'
-import 'bootstrap-toggle/js/bootstrap-toggle'
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Header from './components/Header';
+import About from './components/About';
+import Process from './components/Process';
+import ServiceCatalog from './components/ServiceCatalog';
+import Projects from './components/Projects';
+import Technologies from './components/Technologies';
+import Footer from './components/Footer';
+import ReactGA from 'react-ga'
 
-$(document).ready(function () {
+ReactDOM.render(<Header/>, document.getElementById('header'));
+ReactDOM.render(<About/>, document.getElementById('about'));
+ReactDOM.render(<Process/>, document.getElementById('process'));
+ReactDOM.render(<ServiceCatalog/>, document.getElementById('service_catalog'));
+ReactDOM.render(<Projects/>, document.getElementById('projects'));
+ReactDOM.render(<Technologies/>, document.getElementById('technologies'));
+ReactDOM.render(<Footer/>, document.getElementById('footer'));
 
-    $(function () {
-        $('a[href*="#"]:not([href="#"])').click(function () {
-            if (location.pathname.replace(/^\//, '') == this.pathname.replace(/^\//, '') && location.hostname == this.hostname) {
-                var target = $(this.hash);
-                target = target.length ? target : $('[name=' + this.hash.slice(1) + ']');
+ReactGA.initialize('UA-111095351-1');
 
-                if (target.length) {
-                    $('html, body').animate({
-                        scrollTop: target.offset().top
-                    }, 1000, 'easeInOutExpo');
 
-                    if ($(this).parents('.nav-menu').length) {
-                        $('.nav-menu .menu-active').removeClass('menu-active');
-                        $(this).closest('li').addClass('menu-active');
-                    }
-
-                    if ($('body').hasClass('mobile-nav-active')) {
-                        $('body').removeClass('mobile-nav-active');
-                        $('#mobile-nav-toggle i').toggleClass('fa-times fa-bars');
-                        $('#mobile-body-overly').fadeOut();
-                    }
-                    return false;
-                }
-            }
-        });
-    });
-
-    $('.nav-menu').superfish({
-        animation: {opacity: 'show'},
-        speed: 400
-    });
-
-    if ($('#nav-menu-container').length) {
-        var $mobile_nav = $('#nav-menu-container').clone().prop({id: 'mobile-nav'});
-        $mobile_nav.find('> ul').attr({'class': '', 'id': ''});
-        $('body').append($mobile_nav);
-        $('body').prepend('<button type="button" id="mobile-nav-toggle"><i class="fa fa-bars"></i></button>');
-        $('body').append('<div id="mobile-body-overly"></div>');
-        $('#mobile-nav').find('.menu-has-children').prepend('<i class="fa fa-chevron-down"></i>');
-
-        $(document).on('click', '.menu-has-children i', function (e) {
-            $(this).next().toggleClass('menu-item-active');
-            $(this).nextAll('ul').eq(0).slideToggle();
-            $(this).toggleClass("fa-chevron-up fa-chevron-down");
-        });
-
-        $(document).on('click', '#mobile-nav-toggle', function (e) {
-            $('body').toggleClass('mobile-nav-active');
-            $('#mobile-nav-toggle i').toggleClass('fa-times fa-bars');
-            $('#mobile-body-overly').toggle();
-        });
-
-        $(document).click(function (e) {
-            var container = $("#mobile-nav, #mobile-nav-toggle");
-            if (!container.is(e.target) && container.has(e.target).length === 0) {
-                if ($('body').hasClass('mobile-nav-active')) {
-                    $('body').removeClass('mobile-nav-active');
-                    $('#mobile-nav-toggle i').toggleClass('fa-times fa-bars');
-                    $('#mobile-body-overly').fadeOut();
-                }
-            }
-        });
-    } else if ($("#mobile-nav, #mobile-nav-toggle").length) {
-        $("#mobile-nav, #mobile-nav-toggle").hide();
-    }
-
-    $("#header").sticky({topSpacing: 0, zIndex: '50'});
-
-    $('[data-toggle="tooltip"]').tooltip();
-    $('[data-toggle="popover"]').popover();
-
-    $(window).scroll(function () {
-        if ($(this).scrollTop() > 100) {
-            $('.scrolltop').fadeIn();
-            $('.facebook-messenger').fadeIn();
-        } else {
-            $('.scrolltop').fadeOut();
-            $('.facebook-messenger').fadeOut();
-        }
-    });
-
-    $('.scrolltop, #logo a').click(function () {
-        $("html, body").animate({
-            scrollTop: 0
-        }, 1000, 'easeInOutExpo');
-        return false;
-    });
-
-});