+++ assets/js/presentation.js
@@ -1,0 +1,19 @@
+/*
+ * This file is part of the Hiberbee web application.
+ *
+ * (c) Vladyslav Volkov <vlad@hiberbee.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+import * as Reveal from 'reveal.js/js/reveal'
+
+Reveal.initialize({
+    controls: true,
+    progress: true,
+    history: false,
+    keyboard: true,
+    overview: true,
+    center: true,
+    transition: 'default'
+});